requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
OdooRPC==0.10.1
python-dotenv==0.13.0
requests==2.32.3
urllib3==2.2.2
